+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [2.1.3-dev.2] - 2019-10-17
+
+- fix npm audit issues
+- [#2072](https://github.com/teambit/bit/issues/2072) upgrade module-definition package to support React fragments
+
+
 ## [2.1.3-dev.1] - 2019-10-06
 
 - upgrade typescript and typescript parser packages